requirements.txt
psycopg2==2.9.7
django-environ==0.10.0
Pillow==10.0.0
cloudinary==1.34.0
django-cloudinary-storage==0.3.0